rename gimbal and uart folder structure
