+++ source/gimbal/settings.json
