+++ source/uart/__main__.py
@@ -2,7 +2,7 @@
 Uart driver for sending data to embedded board and receiving corresponding echo
 '''
 
-from source.uart_driver.uart import Uart
+from .uart import Uart
 
 uart = Uart()
